chore: move snapshot command into snapshot.ts
